apps/encoder/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pulsefm-encoder"
version = "0.1.0"
description = "PulseFM audio encoding service."
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.104.1",
  "uvicorn[standard]==0.24.0",
  "pydub==0.25.1",
  "pydantic>=2.0,<3.0",
  "google-cloud-firestore",
  "google-cloud-storage",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["pulsefm_encoder*"]
